requirements.txt
crewai[tools]
langchain
pydantic
python-dotenv
requests
beautifulsoup4
pillow
customtkinter
